JavaFileLineReader
This project is based of to do the following:

Create a utility in Java/Python that counts the actual lines of code in a Java source file

Attached are two test files with code that give you the expected return value n the comments. (Save them so you dont lose the ‘new line/Line Feed’ information)

Remember that Java comments are either // to the end of line, or /* to the next */. The block comments do not nest. There may be multiple /* … */ comments on a line. Whitespace includes tabs, spaces, carriage returns, and vertical tabs. Comment start sequences that appear inside Java strings should be ignored.